% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_missing_files.R
\name{download_missing_files}
\alias{download_missing_files}
\title{Download missing files from AQCSV server}
\usage{
download_missing_files(files_to_get, local_dir = "./COR")
}
\arguments{
\item{files_to_get}{Character vector of file paths to download}

\item{local_path}{Character string of local directory to store files}
}
\value{
invisible NULL
}
\description{
Download missing files from AQCSV server
}
\examples{
files_to_get <- c("file1.csv", "file2.csv")
download_missing_files(files_to_get)
}
