% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aqcsv_creation_date.R
\name{get_aqcsv_dates}
\alias{get_aqcsv_dates}
\title{Get observed and created dates from AQCSV file names}
\usage{
get_aqcsv_dates(
  aqcsv_paths,
  file_suffix = "",
  file_sep = ":",
  date_formats = c(observed = "\%Y\%m\%d\%H\%M", created = "\%Y\%m\%d\%H\%M\%S")
)
}
\arguments{
\item{aqcsv_paths}{Character vector of paths to AQCSV files.}

\item{file_suffix}{Character string to remove from ends file names such that the end of the file is the creation date. Defaults to "" (no file suffix to remove).}

\item{file_sep}{Character string used to seperate pieces of the AQCSV file name. Defaults to ":".}

\item{date_formats}{Named character vector of date formats to use for each date in the file name. Defaults to c(observed = "\%Y\%m\%d\%H\%M", created = "\%Y\%m\%d\%H\%M\%S").}
}
\value{
A data.frame with the first column being the file path and the remaining columns being the extracted dates.
}
\description{
Get observed and created dates from AQCSV file names
}
\examples{
aqcsv_path <- "/some/path/FEM20240101_0000_aq.txt:AQ:PurpleAir:AIRNOW:AQCSV:20260106195903"
get_aqcsv_dates(aqcsv_path)
}
