% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_duplicates.R
\name{get_local_duplicates}
\alias{get_local_duplicates}
\title{Get local files that have duplicate observation dates}
\usage{
get_local_duplicates(file_dates)
}
\arguments{
\item{file_dates}{Data frame of file details with columns for file path, data observation date, and file creation date}
}
\value{
Character vector of file paths of all but the newest of local files with duplicates
}
\description{
Get local files that have duplicate observation dates
}
