% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_paths.R
\name{get_local_paths}
\alias{get_local_paths}
\title{Get all local AQCSV files in the specified directories}
\usage{
get_local_paths(local_path, local_dirs)
}
\arguments{
\item{local_path}{Character string of the local directory to search for files}

\item{local_dirs}{Named character vector of directory names within \code{local_path} that store each set of AQCSVs}
}
\value{
List of character vectors of file paths of all local AQCSV files
}
\description{
Get all local AQCSV files in the specified directories
}
