% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outdated_files.R
\name{get_outdated_files}
\alias{get_outdated_files}
\title{Get paths to local files that are out of date relative to the server}
\usage{
get_outdated_files(server_file_dates, local_file_dates)
}
\arguments{
\item{server_file_dates}{Data frame of server file details with columns for file path, data obervation date, and file creation date}

\item{local_file_dates}{Data frame of local file details with columns for file path, data obervation date, and file creation date}
}
\value{
Character vector of file paths of outdated local files
}
\description{
Get paths to local files that are out of date relative to the server
}
