% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_paths.R
\name{get_server_paths}
\alias{get_server_paths}
\title{Get file paths for AQCSV files on the AQmap server}
\usage{
get_server_paths(
  base_url = "https://aqmap.ca/aqmap/data/AQCSV",
  server_dirs = c(RAW = "raw", COR = "cor", FEM = "fem"),
  server_file_lists = c(RAW = "file_list_raw.txt", COR = "file_list_cor.txt", FEM =
    "file_list_fem.txt")
)
}
\arguments{
\item{base_url}{Base URL for server files}

\item{server_dirs}{Directory names for server files}

\item{server_file_lists}{File names for server file lists}
}
\value{
List of character vectors of server file paths for each of \code{server_dirs}
}
\description{
Get file paths for AQCSV files on the AQmap server
}
