% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_files.R
\name{summarise_files}
\alias{summarise_files}
\title{Output information on provided files for logging}
\usage{
summarise_files(file_paths)
}
\arguments{
\item{file_paths}{Character vector (or list of character vectors) of file paths to summarise}
}
\value{
invisible NULL
}
\description{
Output information on provided files for logging
}
\examples{
file_paths <- c("file1.csv", "file2.csv")
summarise_files(file_paths)
}
