% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_aqcsvs.R
\name{sync_aqcsvs}
\alias{sync_aqcsvs}
\title{Sync AQCSV Files Between AQmap Server and ECCC Science Network}
\usage{
sync_aqcsvs(
  local_path = "./data",
  local_dirs = list(RAW = "RAW", COR = "COR", FEM = "FEM")
)
}
\arguments{
\item{local_path}{Character string of the local directory to store the files}

\item{local_dirs}{Named character vector of directory names within \code{local_path} that store each set of AQCSVs}
}
\value{
invisible NULL (messages output to console for logging)
}
\description{
Syncs the AQCSV files produced on the AQmap server with local files.
}
